# Flatten propagation loop structure