--- a/core/include/detray/propagator/propagator.hpp
+++ b/core/include/detray/propagator/propagator.hpp
@@ -185,97 +185,6 @@ struct propagator {
         propagation._heartbeat = true;
     }
 
-    /// Propagate method init: Initialize a propagation state
-    ///
-    /// @param propagation the state of a propagation flow
-    /// @param actor_state_refs tuple containing refences to the actor states
-    ///
-    /// @return The heartbeat at the end of the initialization.
-    ///
-    /// @note If the return value of this function is true, a propagation step
-    /// can be taken afterwards.
-    template <typename actor_states_t>
-    requires concepts::is_state_of<actor_states_t, actor_chain_type>
-        DETRAY_HOST_DEVICE void propagate_init(
-            state &propagation, actor_states_t actor_state_refs) const {
-        auto &navigation = propagation._navigation;
-        auto &stepping = propagation._stepping;
-        auto &context = propagation._context;
-        const auto &track = stepping();
-        assert(!track.is_invalid());
-
-        // Initialize the navigation
-        m_navigator.init(track, navigation, m_cfg.navigation, context);
-        propagation._heartbeat = navigation.is_alive();
-
-        // Run all registered actors/aborters after init
-        run_actors(actor_state_refs, propagation);
-        assert(!track.is_invalid());
-
-        // Find next candidate
-        m_navigator.update(track, navigation, m_cfg.navigation, context);
-        propagation._heartbeat &= navigation.is_alive();
-    }
-
-    /// Propagate method step: Perform a single propagation step.
-    ///
-    /// @param propagation the state of a propagation flow
-    /// @param actor_state_refs tuple containing refences to the actor states
-    ///
-    /// @return The heartbeat at the end of the step.
-    ///
-    /// @note If the return value of this function is true, another step can
-    /// be taken afterwards.
-    template <typename actor_states_t>
-    requires concepts::is_state_of<actor_states_t, actor_chain_type>
-        DETRAY_HOST_DEVICE bool propagate_step(
-            state &propagation, bool is_init,
-            actor_states_t actor_state_refs) const {
-        auto &navigation = propagation._navigation;
-        auto &stepping = propagation._stepping;
-        auto &context = propagation._context;
-        const auto &track = stepping();
-        assert(!track.is_invalid());
-
-        // Set access to the volume material for the stepper
-        auto vol = navigation.get_volume();
-        const material<scalar_type> *vol_mat_ptr =
-            vol.has_material() ? vol.material_parameters(track.pos()) : nullptr;
-
-        // Break automatic step size scaling by the stepper when a surface
-        // was reached and whenever the navigation is (re-)initialized
-        const bool reset_stepsize{navigation.is_on_surface() || is_init};
-        // Take the step
-        propagation._heartbeat &=
-            m_stepper.step(navigation(), stepping, m_cfg.stepping,
-                           reset_stepsize, vol_mat_ptr);
-
-        // Reduce navigation trust level according to stepper update
-        typename stepper_t::policy_type{}(stepping.policy_state(), propagation);
-
-        // Find next candidate
-        is_init =
-            m_navigator.update(track, navigation, m_cfg.navigation, context);
-        propagation._heartbeat &= navigation.is_alive();
-
-        // Run all registered actors/aborters after update
-        run_actors(actor_state_refs, propagation);
-        assert(!track.is_invalid());
-
-        // And check the status
-        is_init |= m_navigator.update(track, navigation, m_cfg.navigation,
-                                      context, false);
-        propagation._heartbeat &= navigation.is_alive();
-
-#if defined(__NO_DEVICE__)
-        if (propagation.do_debug) {
-            inspect(propagation);
-        }
-#endif
-
-        return is_init;
-    }
-
     /// Propagate method: Coordinates the calls of the stepper, navigator and
     /// all registered actors.
     ///
@@ -289,12 +198,73 @@ struct propagator {
             state &propagation,
             actor_states_t actor_state_refs = dtuple<>{}) const {
 
-        propagate_init(propagation, actor_state_refs);
+        auto &navigation = propagation._navigation;
+        auto &stepping = propagation._stepping;
+        auto &context = propagation._context;
+        const auto &track = stepping();
+        assert(!track.is_invalid());
+
+        // Initialize the navigation
+        m_navigator.init(track, navigation, m_cfg.navigation, context);
+        propagation._heartbeat = navigation.is_alive();
+
         bool is_init = true;
 
-        // Run while there is a heartbeat
-        while (propagation.is_alive()) {
-            is_init = propagate_step(propagation, is_init, actor_state_refs);
+        // Run while there is a heartbeat. In order to help the compiler
+        // optimize this, and in order to make the code more GPU-friendly,
+        // this code is run as a flat loop, but this loop has a defined
+        // structure. Indeed, the structure is always to run either the actors
+        // or the stepper (in alternating order) followed by the navigation
+        // update.
+        //
+        // A = actors
+        // N = navigation update
+        // S = propagation step
+        //
+        // ANSNANSNANSNANSNANSNANS...
+        for (unsigned int i = 0; i % 2 == 0 || propagation.is_alive(); ++i) {
+            if (i % 2 == 0) {
+                // Run all registered actors/aborters
+                run_actors(actor_state_refs, propagation);
+                assert(!track.is_invalid());
+            } else {
+                assert(!track.is_invalid());
+
+                // Set access to the volume material for the stepper
+                auto vol = navigation.get_volume();
+                const material<scalar_type> *vol_mat_ptr =
+                    vol.has_material() ? vol.material_parameters(track.pos())
+                                       : nullptr;
+
+                // Break automatic step size scaling by the stepper when a
+                // surface was reached and whenever the navigation is
+                // (re-)initialized
+                const bool reset_stepsize{navigation.is_on_surface() ||
+                                          is_init};
+                // Take the step
+                propagation._heartbeat &=
+                    m_stepper.step(navigation(), stepping, m_cfg.stepping,
+                                   reset_stepsize, vol_mat_ptr);
+
+                // Reduce navigation trust level according to stepper update
+                typename stepper_t::policy_type{}(stepping.policy_state(),
+                                                  propagation);
+
+                if (i > 0) {
+                    is_init = false;
+                }
+            }
+
+            // Find next candidate
+            is_init |= m_navigator.update(track, navigation, m_cfg.navigation,
+                                          context, i % 2 == 1 || i == 0);
+            propagation._heartbeat &= navigation.is_alive();
+
+#if defined(__NO_DEVICE__)
+            if (i % 2 == 0 && i > 0 && propagation.do_debug) {
+                inspect(propagation);
+            }
+#endif
         }
 
         // Pass on the whether the propagation was successful
@@ -326,13 +296,25 @@ struct propagator {
         DETRAY_HOST_DEVICE bool propagate_sync(
             state &propagation, actor_states_t actor_state_refs) const {
 
-        propagate_init(propagation, actor_state_refs);
-        bool is_init = true;
-
         auto &navigation = propagation._navigation;
         auto &stepping = propagation._stepping;
         auto &context = propagation._context;
         const auto &track = stepping();
+        assert(!track.is_invalid());
+
+        // Initialize the navigation
+        m_navigator.init(track, navigation, m_cfg.navigation, context);
+        propagation._heartbeat = navigation.is_alive();
+
+        // Run all registered actors/aborters after init
+        run_actors(actor_state_refs, propagation);
+        assert(!track.is_invalid());
+
+        // Find next candidate
+        m_navigator.update(track, navigation, m_cfg.navigation, context);
+        propagation._heartbeat &= navigation.is_alive();
+
+        bool is_init = true;
 
         while (propagation.is_alive()) {