This commit flattens the propagation loop structure with the primary goal of reducing the number of different call sites to the navigation update method. Currently. we update the navigator in three different places and this is expensive due to divergence. This commit puts all of those calls into one place in a loop, mirroring the same functional structure that was present before.

I measure a significant performance improvement due to this in traccc, and it also greatly reduces the size of the kernels.