@stephenswat 2025-06-20
### Before 

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   51675881 ns     51580750 ns           13 TracksPropagated=5.07285M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   16518001 ns     16488511 ns           42 TracksPropagated=15.8702M/s
```

### After 

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   46664329 ns     46578745 ns           15 TracksPropagated=5.61765M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   15715844 ns     15687933 ns           45 TracksPropagated=16.6802M/s
```

So seems to improve performance by about 5-10%.