@stephenswat 2025-06-26
Updated performance results (with an added clause to force the navigator update to be inlined):

### Before 

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   48564724 ns     48479844 ns           14 TracksPropagated=5.39783M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   15878392 ns     15849681 ns           44 TracksPropagated=16.5095M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   56845985 ns     56744657 ns           12 TracksPropagated=4.61148M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   17001492 ns     16970615 ns           41 TracksPropagated=15.4189M/s
```

### After 

```
------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                    Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------------------------
BM_PROPAGATION_TOY_DETECTOR_W_COV_TRANSPORT_262144_TRACKS/real_time   37458648 ns     37390653 ns           19 TracksPropagated=6.99822M/s
BM_PROPAGATION_TOY_DETECTOR_262144_TRACKS/real_time                   15085580 ns     15058706 ns           46 TracksPropagated=17.3771M/s
BM_PROPAGATION_WIRE_CHAMBER_W_COV_TRANSPORT_262144_TRACKS/real_time   39743765 ns     39671997 ns           18 TracksPropagated=6.59585M/s
BM_PROPAGATION_WIRE_CHAMBER_262144_TRACKS/real_time                   13186369 ns     13162914 ns           53 TracksPropagated=19.8799M/s
```

Thus the performance gain has now gone up to at least 5% and at most 43% (!).